feat(phone-input): promote next component to default
