+++ documentation-site/examples/phone-input/flag-overrides.js
@@ -3,7 +3,6 @@ import * as React from 'react';
 import {
   PhoneInput,
   COUNTRIES,
-  CountrySelectDropdown,
   StyledFlag,
 } from 'baseui/phone-input';
 import type {CountryIsoT} from 'baseui/phone-input';
@@ -32,15 +31,6 @@ export default function Example() {
       overrides={{
         FlagContainer: {
           component: CustomFlag,
-        },
-        CountrySelect: {
-          props: {
-            overrides: {
-              Dropdown: {
-                component: CountrySelectDropdown,
-              },
-            },
-          },
         },
       }}
     />
