+++ documentation-site/pages/components/phone-input.mdx
@@ -16,7 +16,6 @@ import {Table} from 'baseui/table/index.js';
 import {Block} from 'baseui/block/index.js';
 
 import PhoneInputBasic from 'examples/phone-input/basic.js';
-import PhoneInputNext from 'examples/phone-input/basic-next.js';
 import PhoneInputUncontrolled from 'examples/phone-input/uncontrolled.js';
 import PhoneInputLocalized from 'examples/phone-input/localized.js';
 import FlagOverrides from 'examples/phone-input/flag-overrides.js';
@@ -40,12 +39,6 @@ It contains a dropdown which allows the user to choose from a list of countries 
 <Example title="Basic usage" path="phone-input/basic.js">
   <PhoneInputBasic />
 </Example>
-
-<Example title="Updated phone input design" path="phone-input/basic-next.js">
-  <PhoneInputNext />
-</Example>
-
-Starting in version v11, `PhoneInputNext` will replace the current `PhoneInput`, where the country picker is split from the input element.
 
 <Example title="Localized country names" path="phone-input/localized.js">
   <PhoneInputLocalized />
