+++ src/phone-input/index.d.ts
@@ -335,7 +335,6 @@ export interface PhoneInputProps extends InputProps {
   overrides?: PhoneInputOverrides;
 }
 export const PhoneInput: React.FC<PhoneInputProps>;
-export const PhoneInputNext: React.FC<PhoneInputProps>;
 
 export interface PhoneInputLiteProps extends PhoneInputProps {
   countries: {[key: string]: Country};
