+++ src/phone-input/index.js
@@ -8,17 +8,11 @@ LICENSE file in the root directory of this source tree.
 
 export {default as PhoneInput} from './phone-input.js';
 export {default as PhoneInputLite} from './phone-input-lite.js';
-export {default as PhoneInputNext} from './phone-input-next.js';
 export {default as StatefulPhoneInput} from './stateful-phone-input.js';
-export {
-  default as StatefulPhoneInputNext,
-} from './stateful-phone-input-next.js';
 export {
   default as StatefulPhoneInputContainer,
 } from './stateful-phone-input-container.js';
 export {default as CountryPicker} from './country-picker.js';
-export {default as CountrySelect} from './country-select.js';
-export {default as CountrySelectDropdown} from './country-select-dropdown.js';
 export {default as Flag, StyledFlag} from './flag.js';
 export * from './constants.js';
 export type * from './types.js';
