+++ src/phone-input/phone-input-lite.js
@@ -6,21 +6,21 @@ LICENSE file in the root directory of this source tree.
 */
 // @flow
 import React, {useRef} from 'react';
-import {Input as DefaultInput} from '../input/index.js';
-import CountrySelect from './country-select.js';
-import {getOverrides, mergeOverrides} from '../helpers/overrides.js';
+import CountryPicker from './country-picker.js';
 import defaultProps from './default-props.js';
+import {StyledPhoneInputRoot, StyledDialCode} from './styled-components.js';
+import {Input as DefaultInput} from '../input/index.js';
+import {getOverrides, mergeOverrides} from '../helpers/overrides.js';
 import type {LitePropsT} from './types.js';
 
-const {country, ...lightDefaultProps} = defaultProps;
-
-PhoneInputLite.defaultProps = lightDefaultProps;
+PhoneInputLite.defaultProps = defaultProps;
 
 export default function PhoneInputLite(props: LitePropsT) {
   const {
     'aria-label': ariaLabel,
     'aria-labelledby': ariaLabelledBy,
     'aria-describedby': ariaDescribedBy,
+    clearable,
     countries,
     country,
     disabled,
@@ -41,53 +41,75 @@ export default function PhoneInputLite(props: LitePropsT) {
     ...restProps
   } = props;
   const inputRef = useRef(null);
+
+  const baseDialCodeOverride = {
+    component: StyledDialCode,
+    style: ({$theme: {direction, sizing}}) => {
+      const marginDir = direction === 'rtl' ? 'marginRight' : 'marginLeft';
+      return {
+        [marginDir]: sizing.scale600,
+      };
+    },
+    props: {children: country.dialCode},
+  };
+  const mergedDialCodeOverride = mergeOverrides(
+    {DialCode: baseDialCodeOverride},
+    {DialCode: overrides.DialCode || {}},
+  );
+
   const baseOverrides = {
     Input: {
-      style: ({$theme: {sizing}}) => ({paddingLeft: sizing.scale100}),
-    },
-    Before: {
-      component: CountrySelect,
-      props: {
-        countries,
-        country,
-        disabled,
-        error,
-        inputRef,
-        mapIsoToLabel,
-        maxDropdownHeight,
-        maxDropdownWidth,
-        onCountryChange,
-        overrides,
-        positive,
-        required,
-        size,
+      style: ({$theme: {direction, sizing}}) => {
+        const paddingDir = direction === 'rtl' ? 'paddingRight' : 'paddingLeft';
+        return {
+          [paddingDir]: sizing.scale100,
+        };
       },
     },
+    Before: mergedDialCodeOverride.DialCode,
   };
+  const [Root, rootProps] = getOverrides(overrides.Root, StyledPhoneInputRoot);
   const [Input, inputProps] = getOverrides(overrides.Input, DefaultInput);
   // $FlowFixMe
   inputProps.overrides = mergeOverrides(baseOverrides, inputProps.overrides);
   return (
-    <Input
-      aria-label={ariaLabel}
-      aria-labelledby={ariaLabelledBy}
-      aria-describedby={ariaDescribedBy}
-      autoComplete="tel-national"
-      data-baseweb="phone-input"
-      disabled={disabled}
-      error={error}
-      id={id}
-      inputMode="tel"
-      inputRef={inputRef}
-      name={name}
-      onChange={onTextChange}
-      positive={positive}
-      placeholder={placeholder}
-      size={size}
-      type="text"
-      value={text}
-      {...restProps}
-      {...inputProps}
-    />
+    <Root {...rootProps} data-baseweb="phone-input">
+      <CountryPicker
+        country={country}
+        countries={countries}
+        disabled={disabled}
+        error={error}
+        mapIsoToLabel={mapIsoToLabel}
+        maxDropdownHeight={maxDropdownHeight}
+        maxDropdownWidth={maxDropdownWidth}
+        onCountryChange={onCountryChange}
+        overrides={overrides}
+        positive={positive}
+        required={required}
+        size={size}
+      />
+      <Input
+        aria-label={ariaLabel}
+        aria-labelledby={ariaLabelledBy}
+        aria-describedby={ariaDescribedBy}
+        autoComplete="tel-national"
+        data-baseweb="phone-input"
+        disabled={disabled}
+        error={error}
+        id={id}
+        inputMode="tel"
+        inputRef={inputRef}
+        name={name}
+        onChange={onTextChange}
+        positive={positive}
+        placeholder={placeholder}
+        size={size}
+        type="text"
+        value={text}
+        clearable={clearable}
+        {...restProps}
+        {...inputProps}
+      />
+    </Root>
   );
 }
